% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial}
\alias{binomial}
\title{Calculate binomial distribution P-values between all columns in a matrix}
\usage{
binomial(mat)
}
\arguments{
\item{mat}{a matrix of data, with samples in rows and features in columns}
}
\value{
the binomial P-values between non-zero/missing values in each pair of 
  columns
}
\description{
Calculate P-values from the binomial distribution and take their negative
logarithms as an indicator of coexpression, as proposed by Mohammadi et al.
}
