% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.fk.nz.R
\name{cor.fk.nz}
\alias{cor.fk.nz}
\title{Fast calculation of Kendall's tau rank correlation coefficient for paired
non-zero observations only. Adapted from the pcaPP package source.}
\usage{
cor.fk.nz(x, y = NULL)
}
\arguments{
\item{x}{A vector, a matrix or a data frame of data.}

\item{y}{A vector of data.}
}
\value{
The estimated correlation coefficient.
}
\description{
Fast calculation of Kendall's tau rank correlation coefficient for paired
non-zero observations only. Adapted from the pcaPP package source.
}
