% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Calculate the Dice distance between all columns in a matrix}
\usage{
dice(mat)
}
\arguments{
\item{mat}{a matrix of data, with samples in rows and features in columns}
}
\value{
the Dice coefficient between non-zero/missing values in each pair of 
  columns
}
\description{
Calculate the Dice distance between all pairs of columns in a matrix.
The Dice distance is defined as the number of positions in two vectors
at which the corresponding symbols or different. Here, the Dice distance
is applied to the presence/absence of each gene, whereby
measurements that are missing or equal to zero are considered as zeroes and
all other measurements are considered as ones. This is a wrapper for the
philentropy distance function, that converts a numeric matrix to a 
presence/absence matrix.
}
\examples{
mat = matrix(c(1, rep(0, 9), rep(1, 4), rep(0, 6)), ncol = 2)
dice(mat)
mat = cbind(mat, c(0, rep(1, 5), rep(0, 4)))
dice(mat)

}
