% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{hamming}
\alias{hamming}
\title{Calculate the Hamming distance between all columns in a matrix}
\usage{
hamming(mat)
}
\arguments{
\item{mat}{a matrix of data, with samples in rows and features in columns}
}
\value{
the Hamming index between non-zero/missing values in each pair of 
  columns
}
\description{
Calculate the Hamming distance between all pairs of columns in a matrix.
The Hamming distance is defined as the number of positions in two vectors
at which the corresponding symbols or different. Here, the Hamming distance
is applied to the presence/absence of each gene, whereby
measurements that are missing or equal to zero are considered as zeroes and
all other measurements are considered as ones.
}
\details{
The matrix implementation used in \code{dismay} was obtained from:
\url{https://johanndejong.wordpress.com/2015/10/02/faster-hamming-distance-in-r-2/}
}
\examples{
mat = matrix(c(1, rep(0, 9), rep(1, 4), rep(0, 6)), ncol = 2)
hamming(mat)
mat = cbind(mat, c(0, rep(1, 5), rep(0, 4)))
hamming(mat)

}
