% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{is_bounded}
\alias{is_bounded}
\title{Get metrics bounded by the range [-1, 1]}
\usage{
is_bounded(metric)
}
\value{
true if the metric is within the range [-1, 1], false otherwise
}
\description{
Test whether the values for a metric of interest lie within the range
[-1, 1]. This returns true for Pearson/Spearman/Kendall correlations, 
biweight midcorrelation, zero-inflated Kendall's tau, mutual information
(as implemented in WGCNA), the \code{rho_p} measure of proportionality, 
weighted rank correlation, and cosine similarity.
}
