% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_zi.R
\name{kendall_zi}
\alias{kendall_zi}
\title{Calculate the estimator of Kendall's tau for zero-inflated count data
proposed by Pimentel (Stat Prob Lett (2015) 96:61-67) between each pair of 
columns in a matrix.}
\usage{
kendall_zi(mat)
}
\arguments{
\item{a}{matrix of data}
}
\value{
the estimated correlation coefficients
}
\description{
Calculate the estimator of Kendall's tau for zero-inflated count data
proposed by Pimentel (Stat Prob Lett (2015) 96:61-67) between each pair of 
columns in a matrix.
}
