% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Get all measures of distance or similarity implemented in \code{dismay}}
\usage{
metrics()
}
\value{
all measures of distance or similarity implemented in \code{dismay}
}
\description{
Get all measures of distance or similarity implemented in \code{dismay}
}
