% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialcor.R
\name{partialcor}
\alias{partialcor}
\title{Calculate the partial correlation between all columns in a matrix}
\usage{
partialcor(mat)
}
\arguments{
\item{mat}{a matrix of data, with samples in rows and features in columns}
}
\value{
the estimated partial correlation coefficients
}
\description{
Calculate the partial correlation between all pairs of columns in a matrix.
The partial correlation measures the degree of association between two random
variables, with the effect of a set of controlling random variables removed.
Here, a regularized partial correlation is computed using the corpcor R package.
}
