% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_rank.R
\name{wtd_rank}
\alias{wtd_rank}
\title{Implement weighted rank correlation, as described in Jerrold H Zar, 
"Biostatistical Analysis" 5th ed.}
\usage{
wtd_rank(mat)
}
\arguments{
\item{mat}{a matrix of data}
}
\value{
the estimated correlation coefficients
}
\description{
Implement weighted rank correlation, as described in Jerrold H Zar, 
"Biostatistical Analysis" 5th ed.
}
